Moved POST-request to PostMessage-component
