+++ src/App.jsx
@@ -4,8 +4,6 @@ import { Thoughts } from "./components/Thoughts/Thoughts"
 
 export const App = () => {
   const [thoughts, setThoughts] = useState([])
-  const [newThought, setNewThought] = useState("")
-  //const [submit, setSubmit] = useState(false);
 
   //Fetch existing thoughts on start
   useEffect (() => {
@@ -20,41 +18,25 @@ export const App = () => {
   }, []
   )
 
-  //Set NewThought to the value of inputted message
-  const handleNewMessage = (event) => {
-    setNewThought(event.target.value)
-  }
-
-  //Post a new thought when submit-button is pressed
-  const handlePostSubmit = (event) => {
-        event.preventDefault()
-        fetch("https://happy-thoughts-ux7hkzgmwa-uc.a.run.app/thoughts", {
-        method: "POST",
-        body: JSON.stringify({
-          message: newThought,
-        }),
-        headers: { "Content-Type": "application/json" },
-      })
-        .then((response) => response.json())
-        .then((json) => {
-          setThoughts((previousThoughts) => [json, ...previousThoughts])
-        })
+  //Using json from PostMessage to show new thought
+  const showNewPost = (json) => {
+    setThoughts((previousThoughts) => [json, ...previousThoughts]);
   }
     
   return (
     <div className="app-field">
       <h1>Happy thoughts!</h1>
-      <PostMessage onPostSubmit={handlePostSubmit} onNewMessage={handleNewMessage} newThought={newThought}/>
+      <PostMessage showNewPost={showNewPost}/>
       <div className="thoughts-section">
-       {thoughts.map(thought => (
-        <Thoughts 
-          key={thought._id} 
-          message={thought.message}
-          likes={thought.hearts}
-          time={thought.createdAt}
+        {thoughts.map((thought) => (
+          <Thoughts
+            key={thought._id}
+            message={thought.message}
+            likes={thought.hearts}
+            time={thought.createdAt}
           />
-       ))}
+        ))}
       </div>
     </div>
-  )
+  );
 }
