+++ src/components/Thoughts/TimeStamp.jsx
@@ -29,3 +29,5 @@ export const TimeStamp = ({ time }) => {
 TimeStamp.propTypes = {
   time: PropTypes.string,
 };
+
+//https://date-fns.org/
